# [DSD][Origin Detection] Allow DSD clients to change tags cardinality