--- a/releasenotes/notes/control-dsd-cardinality-via-tags-c0f8b08f424deeda.yaml
+++ b/releasenotes/notes/control-dsd-cardinality-via-tags-c0f8b08f424deeda.yaml
@@ -1,0 +1,5 @@
+---
+features:
+  - |
+    Dogstatsd clients can now choose the cardinality of tags added by origin detection per metrics
+    via the tag 'dd.internal.card' ("low", "orch", "high").