### What does this PR do?

Dogstatsd clients can now choose the cardinality of tags added by origin detection via the tag `dd.internal.card` (low, orch, high).

### Motivation

Support a dynamic cardinality config

### Describe your test plan

`echo -n "dsd.testing.<CARD>:1|g|#foo:bar,dd.internal.entity_id:<Pod UID>,dd.internal.card:<CARD>" | nc -u -w1 <Agent IP> 8125`

`<CARD>` can be `low`, `orch` or `high`