@ahmed-mez 2021-03-04
Thank you @olivielpeau for the review! Here is the benchmark results

Before

```
BenchmarkExtractTagsMetadata
BenchmarkExtractTagsMetadata/20-tags
BenchmarkExtractTagsMetadata/20-tags-8         	31696617	        35.5 ns/op	       0 B/op	       0 allocs/op
BenchmarkExtractTagsMetadata/40-tags
BenchmarkExtractTagsMetadata/40-tags-8         	25163206	        47.5 ns/op	       0 B/op	       0 allocs/op
BenchmarkExtractTagsMetadata/60-tags
BenchmarkExtractTagsMetadata/60-tags-8         	20807089	        64.4 ns/op	       0 B/op	       0 allocs/op
BenchmarkExtractTagsMetadata/80-tags
BenchmarkExtractTagsMetadata/80-tags-8         	13863537	        88.0 ns/op	       0 B/op	       0 allocs/op
BenchmarkExtractTagsMetadata/100-tags
BenchmarkExtractTagsMetadata/100-tags-8        	11995491	       102 ns/op	       0 B/op	       0 allocs/op
BenchmarkExtractTagsMetadata/120-tags
BenchmarkExtractTagsMetadata/120-tags-8        	10091775	       118 ns/op	       0 B/op	       0 allocs/op
BenchmarkExtractTagsMetadata/140-tags
BenchmarkExtractTagsMetadata/140-tags-8        	 9281592	       135 ns/op	       0 B/op	       0 allocs/op
BenchmarkExtractTagsMetadata/160-tags
BenchmarkExtractTagsMetadata/160-tags-8        	 8267480	       148 ns/op	       0 B/op	       0 allocs/op
BenchmarkExtractTagsMetadata/180-tags
BenchmarkExtractTagsMetadata/180-tags-8        	 7227134	       163 ns/op	       0 B/op	       0 allocs/op
BenchmarkExtractTagsMetadata/200-tags
BenchmarkExtractTagsMetadata/200-tags-8        	 6457825	       186 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/DataDog/datadog-agent/pkg/dogstatsd	13.340s
```

After
```
BenchmarkExtractTagsMetadata
BenchmarkExtractTagsMetadata/20-tags
BenchmarkExtractTagsMetadata/20-tags-8         	20539608	        63.4 ns/op	       0 B/op	       0 allocs/op
BenchmarkExtractTagsMetadata/40-tags
BenchmarkExtractTagsMetadata/40-tags-8         	25886548	        46.3 ns/op	       0 B/op	       0 allocs/op
BenchmarkExtractTagsMetadata/60-tags
BenchmarkExtractTagsMetadata/60-tags-8         	16819836	        71.2 ns/op	       0 B/op	       0 allocs/op
BenchmarkExtractTagsMetadata/80-tags
BenchmarkExtractTagsMetadata/80-tags-8         	14505698	        85.8 ns/op	       0 B/op	       0 allocs/op
BenchmarkExtractTagsMetadata/100-tags
BenchmarkExtractTagsMetadata/100-tags-8        	12084690	        97.3 ns/op	       0 B/op	       0 allocs/op
BenchmarkExtractTagsMetadata/120-tags
BenchmarkExtractTagsMetadata/120-tags-8        	10756652	       113 ns/op	       0 B/op	       0 allocs/op
BenchmarkExtractTagsMetadata/140-tags
BenchmarkExtractTagsMetadata/140-tags-8        	 9584504	       129 ns/op	       0 B/op	       0 allocs/op
BenchmarkExtractTagsMetadata/160-tags
BenchmarkExtractTagsMetadata/160-tags-8        	 8116441	       147 ns/op	       0 B/op	       0 allocs/op
BenchmarkExtractTagsMetadata/180-tags
BenchmarkExtractTagsMetadata/180-tags-8        	 7447725	       160 ns/op	       0 B/op	       0 allocs/op
BenchmarkExtractTagsMetadata/200-tags
BenchmarkExtractTagsMetadata/200-tags-8        	 6847184	       161 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/DataDog/datadog-agent/pkg/dogstatsd	15.994s
```